# adding parallel call and async invocation for AMS