# PR Details

## **Thank you for your contribution. Before submitting this PR, please include:**
### Id of the task, bug, story or other reference

### Description
_Include a description of the problem to be solved_

#### Main problem
Performance time for initialization are over 3 seconds, this is mainly due to AMS Client and sequencial calls when loding the componets, which doesnt have a dependency

## Solution Proposed

_Detail what is the solution proposed, include links to design document if required or any other document required to support the solution_

![image](https://github.com/user-attachments/assets/88e9cdf3-c249-4f00-a907-c5a104445bc8)

#### how to use it

```
const initializeOptionalParams: InitializeOptionalParams = {
        getLiveChatConfigOptionalParams: {
            useParallelLoad: true
        }
    };

 await chatSDK.initialize(initializeOptionalParams);
```

The component will have the option to load in parallel or in sequence, for those who doesnt want to change or load time is not a priority, but instead want all components fully loaded.

#### Parallel Load
Load AMS in the background and the rest of the components in parallel, with this change, load times are reduced to the range of 400ms to 700 ms, AMS component may not be fully loaded , but it contains a state machine that handles the states to ensure the component is ready when needed.

#### Sequencial Load
is the legacy implementation, for those that need the component fully loaded before start and performance is not an issue.
This will also serve as switch in case of a corner case no detected and its needed a fallback to legacy from paralllel to sequencial.

#### Other changes

- Fix multiple unit tests with false positives
- Add separate unit tests for parallel and sequential
- update pipeline to use node20


### Acceptance criteria

_Define what are the conditions to consider the PR has achieved the intended goal_

## Test cases and evidence

_Include what tests cases were considered, any evidence of testing for future references, to identify any corner cases, etc_

### OC-LCW Parallel load(~600ms)
<img width="794" alt="image" src="https://github.com/user-attachments/assets/889577fa-7fc4-4edb-9d05-d6d0364450e9">

### OC-LCW Sequential load (~4000ms)
<img width="776" alt="image" src="https://github.com/user-attachments/assets/08436999-4164-4eff-8aea-573ed7b65979">

### OOB Parallel load (~400ms)
![image](https://github.com/user-attachments/assets/734aa1c5-aebd-4526-b454-5995a8ca8232)


### Sanity Tests

- [ ] You have tested all changes in Popout mode
- [ ] You have tested all changes in cross browsers i.e Edge, Chrome, Firefox, Safari and mobile devices(iOS and Android)
- [ ] Your changes are included in the [CHANGELOG](../CHANGE_LOG.md)

## A11y

- [ ] You have run the [Automated Accessibility Check](https://accessibilityinsights.io/docs/en/windows/getstarted/automatedchecks) and have no reported issues

***Please provide justification if any of the validations has been skipped.***